+++ src/providers/seventv/eventapi/Dispatch.hpp
@@ -4,11 +4,15 @@
 
 #pragma once
 
+#include "messages/Emote.hpp"
 #include "providers/seventv/eventapi/Subscription.hpp"
 #include "providers/seventv/SeventvCosmetics.hpp"
 
 #include <QJsonObject>
 #include <QString>
+#include <QVarLengthArray>
+
+#include <span>
 
 namespace chatterino::seventv::eventapi {
 
@@ -81,12 +85,23 @@ struct CosmeticCreateDispatch {
     bool validate() const;
 };
 
+struct TwitchUser {
+    TwitchUser(const QJsonObject &connection);
+    QString id;
+    QString userName;
+};
+
+struct KickUser {
+    KickUser(const QJsonObject &connection);
+
+    uint64_t id;
+    QString userName;
+};
+using User = std::variant<TwitchUser, KickUser>;
+
 struct EntitlementCreateDeleteDispatch {
-    /** id of the user */
-    QString twitchUserID;
-    QString twitchUserName;
-    uint64_t kickUserID;
-    QString kickUserName;
+    QString seventvUsername;
+    QVarLengthArray<User, 1> connections;
     /** id of the entitlement */
     QString refID;
     CosmeticKind kind;
@@ -103,6 +118,11 @@ struct EmoteSetCreateDispatch {
     EmoteSetCreateDispatch(const QJsonObject &emoteSet);
 
     bool validate() const;
+};
+
+struct PersonalEmoteSetAdded {
+    std::span<const User> connections;
+    std::shared_ptr<const EmoteMap> emoteSet;
 };
 
 }  // namespace chatterino::seventv::eventapi